feat: handle multi account with same card ids
A single person might have multiple emails / users in the database.

They might export their data on 1 and import it on another.
We must have a way of allowing the same cardId to be used by multiple
users.

This also helps with data integrity as it prevents a user from
overriding another user's information if they happen to have the same
cardId.

To solve this, we update every table to have a `userId`.
Primary keys are now composite keys which include the `userId`, ensuring
that information is never shared between users.